Determine how much paint is needed to paint the walls
Painting a Room
File
Paint.java
contains the partial program below, which when complete will calculate the amount of paint needed to paint
the walls of a room of the given length and width. It assumes that the paint covers 350 square feet per gallon.
//***************************************************************
//File: Paint.java
//
//Purpose: Determine how much paint is needed to paint the walls
//of a room given its length, width, and height
//***************************************************************
import java.util.Scanner;
public class Paint
{
public static void main(String[] args)
{
final int COVERAGE = 350; //paint covers 350 sq ft/gal
//declare integers length, width, and height;
//declare double totalSqFt;
//declare double paintNeeded;
//declare and initialize Scanner object
//Prompt for and read in the length of the room
//Prompt for and read in the width of the room
//Prompt for and read in the height of the room
//Compute the total square feet to be painted--think
//about the dimensions of each wall
//Compute the amount of paint needed
//Print the length, width, and height of the room and the
//number of gallons of paint needed.
}
}
Save this file to your directory and do the following:
1.
Fill in the missing statements (the comments tell you where to fill in) so that the program does what it is supposed to.
Compile and run the program and correct any errors.
2.
Suppose the room has doors and windows that don't need painting. Ask the user to enter the number of doors and number
of windows in the room, and adjust the total square feet to be painted accordingly. Assume that each door is 20 square
feet and each window is 15 square feet.